+++ src/main/java/roomescape/auth/service/UserAuthService.java
@@ -3,10 +3,10 @@ package roomescape.auth.service;
 import org.springframework.stereotype.Service;
 import roomescape.auth.entity.User;
 import roomescape.auth.repository.UserRepository;
-import roomescape.auth.service.dto.CheckResponse;
-import roomescape.auth.service.dto.LoginRequest;
-import roomescape.auth.service.dto.LoginResponse;
-import roomescape.auth.service.dto.SignupRequest;
+import roomescape.auth.service.dto.response.CheckResponse;
+import roomescape.auth.service.dto.request.LoginRequest;
+import roomescape.auth.service.dto.response.LoginResponse;
+import roomescape.auth.service.dto.request.SignupRequest;
 import roomescape.exception.badRequest.BadRequestException;
 import roomescape.exception.conflict.UserEmailConflictException;
 import roomescape.exception.notFound.UserNotFoundException;
