+++ src/main/java/roomescape/auth/service/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package roomescape.auth.service.dto;
+package roomescape.auth.service.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotEmpty;
