+++ src/main/java/roomescape/auth/service/dto/response/CheckResponse.java
@@ -1,4 +1,4 @@
-package roomescape.auth.service.dto;
+package roomescape.auth.service.dto.response;
 
 public record CheckResponse(
         String name
