+++ src/main/java/roomescape/auth/service/dto/response/LoginResponse.java
@@ -1,4 +1,4 @@
-package roomescape.auth.service.dto;
+package roomescape.auth.service.dto.response;
 
 public record LoginResponse(
         String token
